requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai>=1.10.0
elevenlabs==0.2.26
moviepy==1.0.3
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
jinja2==3.1.2 